Develop

+++ app/Repositories/TransversalActivityRepository.php
@@ -29,10 +29,19 @@ class TransversalActivityRepository
 
     public function getAll()
     {
-        // $rol_id = $this->getIdRolUserAuth();
+        $rol_id = $this->getIdRolUserAuth();
         $user_id = $this->getIdUserAuth();
 
-        $query = $this->model->query()->where('created_by', $user_id)->orderBy('id', 'DESC');
+        $query = $this->model->query();
+
+        if ($rol_id == config('roles.director_programa')) {
+            $query->where('status_id', config('status.ENR'))
+                ->whereNotIn('created_by', [1,2]);
+        }
+
+        if ($rol_id == config('roles.psicologo')) {
+            $query->where('created_by', $user_id)->orderBy('id', 'DESC');
+        }
 
         $paginate = config('global.paginate');
 
@@ -86,24 +95,39 @@ class TransversalActivityRepository
 
     public function update($request, $id)
     {
-        // $user_id = $this->getIdUserAuth();
+        $rol_id = $this->getIdRolUserAuth();
+        $user_id = $this->getIdUserAuth();
 
-        $transversalActivity = $this->model->findOrFail($id);;
-        $transversalActivity->date_visit = $request['date_visit'];
-        $transversalActivity->nro_assistants = $request['nro_assistants'];
-        $transversalActivity->activity_name = $request['activity_name'];
-        $transversalActivity->objective_activity = $request['objective_activity'];
-        $transversalActivity->scene = $request['scene'];
-        $transversalActivity->meeting_planing = $request['meeting_planing'];
-        $transversalActivity->team_socialization = $request['team_socialization'];
-        $transversalActivity->development_activity = $request['development_activity'];
-        $transversalActivity->content_network = $request['content_network'];
-        $transversalActivity->municipality_id = $request['municipality_id'];
-        // $transversalActivity->created_by = $user_id;
+        $transversalActivity = $this->model->findOrFail($id);
+
+        /* CAMBIAMOS EL ESTADO SEGUN EL ROL ASIGNADO A REVISAR */
+        if ($rol_id == config('roles.director_programa')) {
+            $transversalActivity->reviewed_by = $user_id;
+            $transversalActivity->status_id = $request['status_id'];
+            $transversalActivity->reject_message = $request['reject_message'];
+        }
+
+        /* CAMBIAMOS EL ESTADO SI ESTA RECHAZADO*/
+        if ($request['status_id'] == config('status.REC') && $user_id == $transversalActivity->created_by) {
+            $transversalActivity->status_id = config('status.ENR');
+            $transversalActivity->reject_message = $request['reject_message'];
+        }
+
+        if ($user_id == $transversalActivity->created_by) {
+            $transversalActivity->date_visit = $request['date_visit'];
+            $transversalActivity->nro_assistants = $request['nro_assistants'];
+            $transversalActivity->activity_name = $request['activity_name'];
+            $transversalActivity->objective_activity = $request['objective_activity'];
+            $transversalActivity->scene = $request['scene'];
+            $transversalActivity->meeting_planing = $request['meeting_planing'];
+            $transversalActivity->team_socialization = $request['team_socialization'];
+            $transversalActivity->development_activity = $request['development_activity'];
+            $transversalActivity->content_network = $request['content_network'];
+            $transversalActivity->municipality_id = $request['municipality_id'];
+        }
         $save = $transversalActivity->save();
-
         // /* SUBIMOS EL ARCHIVO */
-        if ($save) {
+        if ($save && $rol_id != config('roles.director_programa')) {
             $this->updateAllFiles($request, $transversalActivity->id);
         }
 
@@ -139,7 +163,7 @@ class TransversalActivityRepository
             'development_activity' => 'bail|required',
             'content_network' => 'bail|required',
             'municipality_id' => 'bail|required',
-            'file' => 'bail|required',
+            //'file' => $method != 'update' ? 'bail|required|mimes:application/pdf,pdf,png,webp,jpg,jpeg|max:' . (500 * 1049000000) : 'bail',
         ];
 
         $messages = [
@@ -181,21 +205,21 @@ class TransversalActivityRepository
                 foreach ($request->file('file') as $file) {
                     $name = strtolower(Str::random(10));
                     $image = Image::make($file);
-                    $image->encode('webp', 90);
+                    $image->encode('jpg', 90);
                     // Create folder directory and save
                     Storage::disk('public')->makeDirectory('transversal_activities/' . $id);
-                    $image->save($path . $name . '.webp');
+                    $image->save($path . $name . '.jpg');
                     // Save url image for load in database
-                    $url = "transversal_activities/{$id}/{$name}.webp";
+                    $url = "transversal_activities/{$id}/{$name}.jpg";
                     // Save in database with relation
                     $this->modelEvidence->create([
-                        'name' => "Evidencia Actividad Transversal $id",
+                        'model' => "Evidencia Actividad Transversal $id",
                         'path' => $url,
                         'transversal_id' => $id,
                     ]);
                 }
                 return [
-                    'response' => ['status' => true, 'name' => $name, 'message' => 'Se ha guardado con éxito']
+                    'response' => ['status' => true, 'name' => 'transversal_activities', 'message' => 'Se ha guardado con éxito']
                 ];
             } catch (\Exception $ex) {
                 return [
@@ -209,12 +233,16 @@ class TransversalActivityRepository
     public function updateAllFiles($request, $id) {
 
         try {
+            $files = $request->file('file');
             $evidences = $this->modelEvidence->where('transversal_id', $id)->get();
-            foreach ($evidences as $evidence) {
-                Storage::disk('public')->delete($evidence->path);
+            if (!empty($files)) {
+                foreach ($evidences as $evidence) {
+                    Storage::disk('public')->delete($evidence->path);
+                    $evidence->delete();
+                }
+                $response = $this->uploadAll($request, $id);
+                return $response;
             }
-            $response = $this->uploadAll($request, $id);
-            return $response;
         } catch (Exception $ex) {
             return [
                 'response' => ['success' => false, 'payload' => $ex->getMessage()]
